refine(match-version.js): Don't require 'refs/tags/' in the prefix
I believe there was an error in the last test as it did not have
'gitRef' parameter did not start with 'refs/tags/' like in all the
other tests.  This inconsistency lead to a passing test.

When the 'gitRef' parameter of the last test starts with 'refs/tags/'
the test fails, which leads me to believe 'match-version.js' is not
behaving as intended.

I don't think we need the calling code to include 'refs/tags/' in
the TAG_PREFIX, although for backward compatibility, it should
be allowed.